# backend/openapi/openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "description" : "Contrato HTTP do CRM PNP. Tenant é sempre derivado da credencial.",
    "title" : "CRM PNP API",
    "version" : "1.0.0"
  },
  "servers" : [ {
    "url" : "http://localhost",
    "description" : "Generated server url"
  } ],
  "paths" : {
...
    "/api/automacoes" : {
      "get" : {
        "operationId" : "listar_4",
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/DefinicaoResumo"
                  }
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErroResponse"
                }
              }
            },
            "description" : "Requisição ou JSON inválido"
          },
          "401" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErroResponse"
                }
              }
            },
            "description" : "Autenticação ausente ou expirada"
          },
          "403" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErroResponse"
                }
              }
            },
            "description" : "Acesso negado"
          },
          "404" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErroResponse"
                }
              }
            },
            "description" : "Recurso não encontrado"
          },
          "409" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErroResponse"
                }
              }
            },
            "description" : "Conflito de estado ou idempotência"
          },
          "422" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErroResponse"
                }
              }
            },
            "description" : "Regra de negócio ou validação semântica"
          },
          "429" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErroResponse"
                }
              }
            },
            "description" : "Limite de requisições excedido"
          },
          "500" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErroResponse"
                }
              }
            },
            "description" : "Falha interna com detalhe seguro"
          }
        },
        "tags" : [ "automation-controller" ]
      },
      "post" : {
        "operationId" : "criar_5",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/DefinicaoRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/DefinicaoResumo"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErroResponse"
                }
              }
            },
            "description" : "Requisição ou JSON inválido"
          },
          "401" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErroResponse"
                }
              }
            },
            "description" : "Autenticação ausente ou expirada"
          },
          "403" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErroResponse"
                }
              }
            },
            "description" : "Acesso negado"
          },
          "404" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErroResponse"
                }
              }
            },
            "description" : "Recurso não encontrado"
          },
          "409" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErroResponse"
                }
              }
            },
            "description" : "Conflito de estado ou idempotência"
          },
          "422" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErroResponse"
                }
              }
            },
            "description" : "Regra de negócio ou validação semântica"
          },
          "429" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErroResponse"
                }
              }
            },
            "description" : "Limite de requisições excedido"
          },
          "500" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErroResponse"
                }
              }
            },
            "description" : "Falha interna com detalhe seguro"
          }
        },
        "tags" : [ "automation-controller" ]
      }
    },
    "/api/automacoes/execucoes" : {
      "get" : {
        "operationId" : "listarExecucoes",
        "parameters" : [ {
          "in" : "query",
          "name" : "definicaoId",
          "required" : false,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/ExecucaoResumo"
                  }
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErroResponse"
                }
              }
            },
            "description" : "Requisição ou JSON inválido"
          },
          "401" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErroResponse"
                }
              }
            },
            "description" : "Autenticação ausente ou expirada"
          },
          "403" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErroResponse"
                }
              }
            },
            "description" : "Acesso negado"
          },
          "404" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErroResponse"
                }
              }
            },
            "description" : "Recurso não encontrado"
          },
          "409" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErroResponse"
                }
              }
            },
            "description" : "Conflito de estado ou idempotência"
          },
          "422" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErroResponse"
                }
              }
            },
            "description" : "Regra de negócio ou validação semântica"
          },
          "429" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErroResponse"
                }
              }
            },
            "description" : "Limite de requisições excedido"
          },
          "500" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErroResponse"
                }
              }
            },
            "description" : "Falha interna com detalhe seguro"
          }
        },
        "tags" : [ "automation-controller" ]
      }
    },
    "/api/automacoes/execucoes/{id}" : {
      "get" : {
        "operationId" : "buscarExecucao",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ExecucaoResumo"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErroResponse"
                }
              }
            },
            "description" : "Requisição ou JSON inválido"
          },
          "401" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErroResponse"
                }
              }
            },
            "description" : "Autenticação ausente ou expirada"
          },
          "403" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErroResponse"
                }
              }
            },
            "description" : "Acesso negado"
          },
          "404" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErroResponse"
                }
              }
            },
            "description" : "Recurso não encontrado"
          },
          "409" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErroResponse"
                }
              }
            },
            "description" : "Conflito de estado ou idempotência"
          },
          "422" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErroResponse"
                }
              }
            },
            "description" : "Regra de negócio ou validação semântica"
          },
          "429" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErroResponse"
                }
              }
            },
            "description" : "Limite de requisições excedido"
          },
          "500" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErroResponse"
                }
              }
            },
            "description" : "Falha interna com detalhe seguro"
          }
        },
        "tags" : [ "automation-controller" ]
      }
    },
    "/api/automacoes/execucoes/{id}/cancelar" : {
      "post" : {
        "operationId" : "cancelar",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ExecucaoResumo"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErroResponse"
                }
              }
            },
            "description" : "Requisição ou JSON inválido"
          },
          "401" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErroResponse"
                }
              }
            },
            "description" : "Autenticação ausente ou expirada"
          },
          "403" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErroResponse"
                }
              }
            },
            "description" : "Acesso negado"
          },
          "404" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErroResponse"
                }
              }
            },
            "description" : "Recurso não encontrado"
          },
          "409" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErroResponse"
                }
              }
            },
            "description" : "Conflito de estado ou idempotência"
          },
          "422" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErroResponse"
                }
              }
            },
            "description" : "Regra de negócio ou validação semântica"
          },
          "429" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErroResponse"
                }
              }
            },
            "description" : "Limite de requisições excedido"
          },
          "500" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErroResponse"
                }
              }
            },
            "description" : "Falha interna com detalhe seguro"
          }
        },
        "tags" : [ "automation-controller" ]
      }
    },
    "/api/automacoes/execucoes/{id}/pausar" : {
      "post" : {
        "operationId" : "pausarExecucao",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ExecucaoResumo"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErroResponse"
                }
              }
            },
            "description" : "Requisição ou JSON inválido"
          },
          "401" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErroResponse"
                }
              }
            },
            "description" : "Autenticação ausente ou expirada"
          },
          "403" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErroResponse"
                }
              }
            },
            "description" : "Acesso negado"
          },
          "404" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErroResponse"
                }
              }
            },
            "description" : "Recurso não encontrado"
          },
          "409" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErroResponse"
                }
              }
            },
            "description" : "Conflito de estado ou idempotência"
          },
          "422" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErroResponse"
                }
              }
            },
            "description" : "Regra de negócio ou validação semântica"
          },
          "429" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErroResponse"
                }
              }
            },
            "description" : "Limite de requisições excedido"
          },
          "500" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErroResponse"
                }
              }
            },
            "description" : "Falha interna com detalhe seguro"
          }
        },
        "tags" : [ "automation-controller" ]
      }
    },
    "/api/automacoes/execucoes/{id}/retomar" : {
      "post" : {
        "operationId" : "retomar",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ExecucaoResumo"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErroResponse"
                }
              }
            },
            "description" : "Requisição ou JSON inválido"
          },
          "401" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErroResponse"
                }
              }
            },
            "description" : "Autenticação ausente ou expirada"
          },
          "403" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErroResponse"
                }
              }
            },
            "description" : "Acesso negado"
          },
          "404" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErroResponse"
                }
              }
            },
            "description" : "Recurso não encontrado"
          },
          "409" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErroResponse"
                }
              }
            },
            "description" : "Conflito de estado ou idempotência"
          },
          "422" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErroResponse"
                }
              }
            },
            "description" : "Regra de negócio ou validação semântica"
          },
          "429" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErroResponse"
                }
              }
            },
            "description" : "Limite de requisições excedido"
          },
          "500" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErroResponse"
                }
              }
            },
            "description" : "Falha interna com detalhe seguro"
          }
        },
        "tags" : [ "automation-controller" ]
      }
    },
    "/api/automacoes/execucoes/{id}/transicoes" : {
      "get" : {
        "operationId" : "transicoes",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/TransicaoResumo"
                  }
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErroResponse"
                }
              }
            },
            "description" : "Requisição ou JSON inválido"
          },
          "401" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErroResponse"
                }
              }
            },
            "description" : "Autenticação ausente ou expirada"
          },
          "403" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErroResponse"
                }
              }
            },
            "description" : "Acesso negado"
          },
          "404" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErroResponse"
                }
              }
            },
            "description" : "Recurso não encontrado"
          },
          "409" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErroResponse"
                }
              }
            },
            "description" : "Conflito de estado ou idempotência"
          },
          "422" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErroResponse"
                }
              }
            },
            "description" : "Regra de negócio ou validação semântica"
          },
          "429" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErroResponse"
                }
              }
            },
            "description" : "Limite de requisições excedido"
          },
          "500" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErroResponse"
                }
              }
            },
            "description" : "Falha interna com detalhe seguro"
          }
        },
        "tags" : [ "automation-controller" ]
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "DefinicaoRequest" : {
        "type" : "object",
        "properties" : {
          "codigo" : {
            "type" : "string",
            "minLength" : 1,
            "pattern" : "^[a-z][a-z0-9_-]{2,79}$"
          },
          "especificacao" : {
            "$ref" : "#/components/schemas/EspecificacaoRequest"
          },
          "limiteConcorrencia" : {
            "type" : "integer",
            "format" : "int32",
            "maximum" : 20,
            "minimum" : 1
          },
          "nome" : {
            "type" : "string",
            "maxLength" : 200,
            "minLength" : 0
          }
        },
        "required" : [ "codigo", "especificacao", "nome" ]
      },
      "DefinicaoResumo" : {
        "type" : "object",
        "properties" : {
          "atualizadaEm" : {
            "type" : "string",
            "format" : "date-time"
          },
          "codigo" : {
            "type" : "string"
          },
          "criadaEm" : {
            "type" : "string",
            "format" : "date-time"
          },
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "limiteConcorrencia" : {
            "type" : "integer",
            "format" : "int32"
          },
          "nome" : {
            "type" : "string"
          },
          "status" : {
            "type" : "string"
          },
          "versaoAtiva" : {
            "type" : "integer",
            "format" : "int32"
          },
          "versaoAtual" : {
            "type" : "integer",
            "format" : "int32"
          }
        }
      },
...
      "ErroCampo" : {
        "properties" : {
          "campo" : {
            "type" : "string"
          },
          "mensagem" : {
            "type" : "string"
          }
        }
      },
      "ErroResponse" : {
        "properties" : {
          "campos" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/ErroCampo"
            }
          },
          "codigo" : {
            "type" : "string"
          },
          "correlacaoId" : {
            "type" : "string"
          },
          "detail" : {
            "type" : "string"
          },
          "instance" : {
            "type" : "string",
            "format" : "uri"
          },
          "momento" : {
            "type" : "string",
            "format" : "date-time"
          },
          "status" : {
            "type" : "integer",
            "format" : "int32"
          },
          "title" : {
            "type" : "string"
          },
          "type" : {
            "type" : "string",
            "format" : "uri"
          }
        }
      },
...
      "ExecucaoResumo" : {
        "type" : "object",
        "properties" : {
          "codigoFalha" : {
            "type" : "string"
          },
          "criadaEm" : {
            "type" : "string",
            "format" : "date-time"
          },
          "definicaoId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "finalizadaEm" : {
            "type" : "string",
            "format" : "date-time"
          },
          "gatilho" : {
            "type" : "string"
          },
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "iniciadaEm" : {
            "type" : "string",
            "format" : "date-time"
          },
          "limiteEm" : {
            "type" : "string",
            "format" : "date-time"
          },
          "simulacao" : {
            "type" : "boolean"
          },
          "status" : {
            "type" : "string"
          },
          "versaoDefinicao" : {
            "type" : "integer",
            "format" : "int32"
          }
        }
      },
...
    }
  }
}